refactor: move the reference wire shapes into @virtool/contracts
The reference shapes were defined in server/references/data.ts and
re-exported to components through references/types.ts, so the browser's
view of `Reference` depended on a data-layer module's emitted
declarations — a data.ts refactor could break client types. The
camelCase pagination envelope was worse: server/references/data.ts kept
a private copy with a comment explaining it could not import the
client's.

The shapes now live in @virtool/contracts, which neither side reaches
through the other: references.ts for the reference shapes, plus
users.ts (UserNested), tasks.ts (Task) and search.ts (SearchResultV2)
for what they embed. data.ts imports them to annotate its returns and
keeps only what it alone uses; references/types.ts re-exports them, the
pattern already used for roles, permissions and banner colors. The
generic ClonedFrom and ImportedFrom gain a Reference prefix now that
they are package-level names.
