+++ packages/contracts/src/index.ts
@@ -9,7 +9,11 @@ export * from "./indexes";
 export * from "./jobs";
 export * from "./permissions";
 export * from "./permissionsSchema";
+export * from "./references";
 export * from "./samples";
+export * from "./search";
 export * from "./sse";
 export * from "./subtractions";
+export * from "./tasks";
+export * from "./users";
 export * from "./workflowName";
