+++ packages/contracts/src/users.ts
@@ -1,0 +1,8 @@
+/** A user reduced to the fields shown alongside another resource. */
+export type UserNested = {
+	/** The unique identifier */
+	id: number;
+
+	/** The user's handle or username */
+	handle: string;
+};
